requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.3
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
bcrypt==4.0.1
mux_python==5.0.1
httpx==0.28.1